rename withValidationHandler to withFieldValidation
